Refactor ActivityTokens to use a trait that can be shared with other entities
